tests prerendering next9.3
